+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @doodle-engine/core
 
+## 0.0.20
+
+### Patch Changes
+
+- 96d5194: fix missing css
+
+## 0.0.19
+
+### Patch Changes
+
+- 07090d1: feature: asset loader
+
 ## 0.0.18
 
 ### Patch Changes