+++ packages/react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @doodle-engine/react
 
+## 0.0.20
+
+### Patch Changes
+
+- 96d5194: fix missing css
+- Updated dependencies [96d5194]
+  - @doodle-engine/core@0.0.20
+
+## 0.0.19
+
+### Patch Changes
+
+- 07090d1: feature: asset loader
+- Updated dependencies [07090d1]
+  - @doodle-engine/core@0.0.19
+
 ## 0.0.18
 
 ### Patch Changes